// estela-web/src/services/api/generated-api/models/Connection.ts
/* tslint:disable */
/* eslint-disable */
/**
 * estela API v1.0 Documentation
 * estela API Swagger Specification
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Connection
 */
export interface Connection {
    /**
     * Connection ID.
     * @type {number}
     * @memberof Connection
     */
    readonly cid?: number;
    /**
     * Connection name.
     * @type {string}
     * @memberof Connection
     */
    name: string;
    /**
     * Connection type.
     * @type {string}
     * @memberof Connection
     */
    connType: ConnectionConnTypeEnum;
    /**
     * Connection host.
     * @type {string}
     * @memberof Connection
     */
    host?: string;
    /**
     * Connection port.
     * @type {string}
     * @memberof Connection
     */
    port?: string;
    /**
     * Connection login/username.
     * @type {string}
     * @memberof Connection
     */
    login?: string;
    /**
     * Connection password.
     * @type {string}
     * @memberof Connection
     */
    password?: string;
    /**
     * Additional connection configuration.
     * @type {object}
     * @memberof Connection
     */
    extra?: object;
    /**
     * Project this connection belongs to.
     * @type {string}
     * @memberof Connection
     */
    readonly project?: string;
    /**
     * Creation date.
     * @type {Date}
     * @memberof Connection
     */
    readonly created?: Date;
    /**
     * Last update date.
     * @type {Date}
     * @memberof Connection
     */
    readonly updated?: Date;
}

/**
* @export
* @enum {string}
*/
export enum ConnectionConnTypeEnum {
    Database = 'DATABASE',
    S3 = 'S3',
    Ftp = 'FTP',
    Sftp = 'SFTP',
    Http = 'HTTP',
    Webhook = 'WEBHOOK'
}

export function ConnectionFromJSON(json: any): Connection {
    return ConnectionFromJSONTyped(json, false);
}

export function ConnectionFromJSONTyped(json: any, ignoreDiscriminator: boolean): Connection {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'cid': !exists(json, 'cid') ? undefined : json['cid'],
        'name': json['name'],
        'connType': json['conn_type'],
        'host': !exists(json, 'host') ? undefined : json['host'],
        'port': !exists(json, 'port') ? undefined : json['port'],
        'login': !exists(json, 'login') ? undefined : json['login'],
        'password': !exists(json, 'password') ? undefined : json['password'],
        'extra': !exists(json, 'extra') ? undefined : json['extra'],
        'project': !exists(json, 'project') ? undefined : json['project'],
        'created': !exists(json, 'created') ? undefined : (new Date(json['created'])),
        'updated': !exists(json, 'updated') ? undefined : (new Date(json['updated'])),
    };
}

export function ConnectionToJSON(value?: Connection | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'conn_type': value.connType,
        'host': value.host,
        'port': value.port,
        'login': value.login,
        'password': value.password,
        'extra': value.extra,
    };
}